feat: Adjusted way EstablishmentDetais is being rendered. Fixed stuck lists on appointments list.
