--- Client/screens/AppointmentNavigator.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import ScheduleNavigator from './ScheduleNavigator';
-import AppointmentNav from '@navigation/AppointmentNavigator';
-
-export default function AppointmentNavigator() {
-    return (
-        <ScheduleNavigator Nav={AppointmentNav} />
-    );
-}
