--- Client/screens/HomeNavigator.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import HomeNav from '@navigation/HomeNavigator';
-import ScheduleNavigator from './ScheduleNavigator';
-
-export default function HomeNavigator() {
-    return (
-        <ScheduleNavigator Nav={HomeNav} />
-    );
-}
